+++ lib/src/stats/metric_server.dart
@@ -3,21 +3,12 @@ import 'dart:io';
 import '../http/http.dart';
 import 'stats.dart';
 
-/// A variant of an [Angel] server that records performance metrics.
+@deprecated
 class AngelMetrics extends Angel {
   Angel _inner;
-  HttpServer _server;
   StreamSubscription<HttpRequest> _sub;
 
   AngelMetrics() : super() {
-    var zoneBuilder = createZoneForRequest;
-    createZoneForRequest = (request, req, res) async {
-      var spec = await zoneBuilder(request, req, res);
-      return new ZoneSpecification.from(
-        spec,
-      );
-    };
-
     get('/metrics', (req, res) {
       res.contentType = ContentType.HTML;
 
@@ -63,30 +54,7 @@ class AngelMetrics extends Angel {
     });
   }
 
-  factory AngelMetrics.custom(ServerGenerator serverGenerator) {
-    return new AngelMetrics().._inner = new Angel.custom(serverGenerator);
-  }
-
-  @override
-  HttpServer get httpServer => _server ?? super.httpServer;
-
   final AngelMetricsStats stats = new AngelMetricsStats._();
-
-  @override
-  Future<HttpServer> startServer([address, int port]) async {
-    if (_inner == null) return await super.startServer(address, port);
-
-    var host = address ?? InternetAddress.LOOPBACK_IP_V4;
-    _server = await _inner.serverGenerator(host, port ?? 0);
-
-    for (var configurer in startupHooks) {
-      await configure(configurer);
-    }
-
-    optimizeForProduction();
-    _sub = _server.listen(handleRequest);
-    return _server;
-  }
 
   @override
   Future<HttpServer> close() async {
@@ -96,30 +64,10 @@ class AngelMetrics extends Angel {
   }
 
   @override
-  Future<RequestContext> createRequestContext(HttpRequest request) {
-    return stats.createRequestContext
-        .run<RequestContext>(() => super.createRequestContext(request));
-  }
-
-  @override
-  Future<ResponseContext> createResponseContext(HttpResponse response,
-      [RequestContext correspondingRequest]) {
-    return stats.createResponseContext.run<ResponseContext>(
-        () => super.createResponseContext(response, correspondingRequest));
-  }
-
-  @override
   Iterable<RoutingResult> resolveAll(String absolute, String relative,
       {String method: 'GET', bool strip: true}) {
     return stats.resolveAll.run(() =>
-        super.resolveAll(absolute, relative, method: method, strip: strip));
-  }
-
-  @override
-  Future handleRequest(HttpRequest request) {
-    return stats.handleRequest.run(() async {
-      await super.handleRequest(request);
-    });
+        _inner.resolveAll(absolute, relative, method: method, strip: strip));
   }
 
   @override
@@ -140,37 +88,22 @@ class AngelMetrics extends Angel {
       Function handler, RequestContext req, ResponseContext res) {
     return stats.runContained.run(() => super.runContained(handler, req, res));
   }
-
-  @override
-  Future sendResponse(
-      HttpRequest request, RequestContext req, ResponseContext res,
-      {bool ignoreFinalizers: false}) {
-    return stats.sendResponse.run(() => super.sendResponse(request, req, res));
-  }
 }
 
 class AngelMetricsStats {
   AngelMetricsStats._() {
     all = [
-      createRequestContext,
-      createResponseContext,
       resolveAll,
       executeHandler,
       getHandlerResult,
       runContained,
-      sendResponse,
-      handleRequest,
     ];
   }
 
-  final Stats createRequestContext = new Stats('createRequestContext');
-  final Stats createResponseContext = new Stats('createResponseContext');
   final Stats resolveAll = new Stats('resolveAll');
-  final Stats handleRequest = new Stats('handleRequest');
   final Stats executeHandler = new Stats('executeHandler');
   final Stats getHandlerResult = new Stats('getHandlerResult');
   final Stats runContained = new Stats('runContained');
-  final Stats sendResponse = new Stats('sendResponse');
 
   List<Stats> all;